fixed ci tests

+++ test/forge/TestContract.t.sol
@@ -16,6 +16,7 @@ pragma solidity 0.8.19;
 import '../../lib/forge-std/src/Test.sol';
 
 contract TestContract is Test {
+
   using stdStorage for StdStorage;
 
   Storage test;
@@ -52,39 +53,29 @@ contract TestContract is Test {
   // If targeting a mapping, you have to pass in the keys necessary to perform the find
   // i.e.:
   function testFindMapping() public {
-    uint256 slot = stdstore
-      .target(address(test))
-      .sig(test.map_addr.selector)
-      .with_key(address(this))
-      .find();
+    uint256 slot = stdstore.target(address(test)).sig(test.map_addr.selector).with_key(address(this)).find();
 
     // in the `Storage` constructor, we wrote that this address' value was 1 in the map
     // so when we load the slot, we expect it to be 1
-    assertEq(uint(vm.load(address(test), bytes32(slot))), 1);
+    assertEq(uint256(vm.load(address(test), bytes32(slot))), 1);
   }
 
   // If the target is a struct, you can specify the field depth:
   function testFindStruct() public {
     // NOTE: see the depth parameter - 0 means 0th field, 1 means 1st field, etc.
-    uint256 slot_for_a_field = stdstore
-      .target(address(test))
-      .sig(test.basicStruct.selector)
-      .depth(0)
-      .find();
+    uint256 slot_for_a_field = stdstore.target(address(test)).sig(test.basicStruct.selector).depth(0).find();
 
-    uint256 slot_for_b_field = stdstore
-      .target(address(test))
-      .sig(test.basicStruct.selector)
-      .depth(1)
-      .find();
+    uint256 slot_for_b_field = stdstore.target(address(test)).sig(test.basicStruct.selector).depth(1).find();
 
-    assertEq(uint(vm.load(address(test), bytes32(slot_for_a_field))), 1);
-    assertEq(uint(vm.load(address(test), bytes32(slot_for_b_field))), 2);
+    assertEq(uint256(vm.load(address(test), bytes32(slot_for_a_field))), 1);
+    assertEq(uint256(vm.load(address(test), bytes32(slot_for_b_field))), 2);
   }
+
 }
 
 // A complex storage contract
 contract Storage {
+
   struct UnpackedStruct {
     uint256 a;
     uint256 b;
@@ -100,7 +91,7 @@ contract Storage {
   mapping(address => UnpackedStruct) public map_struct;
   mapping(address => mapping(address => uint256)) public deep_map;
   mapping(address => mapping(address => UnpackedStruct)) public deep_map_struct;
-  UnpackedStruct public basicStruct = UnpackedStruct({a: 1, b: 2});
+  UnpackedStruct public basicStruct = UnpackedStruct({ a: 1, b: 2 });
 
   function hidden() public view returns (bytes32 t) {
     // an extremely hidden storage slot
@@ -109,4 +100,5 @@ contract Storage {
       t := sload(slot)
     }
   }
+
 }
